+++ src/Resources/contao/languages/en/tl_page.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_page'];
+
+/*
+ * Fields
+ */
+$lang['categoriesParam'][0] = 'Custom category parameter';
+$lang['categoriesParam'][1] = 'Set a custom category GET parameter. Default is "category".';
+
+/*
+ * Legends
+ */
+$lang['categories_legend'] = 'Category settings';